# apps/control-plane/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenCrane Control Plane API",
    "version": "1.0.0",
    "description": "Multi-tenant AI agent platform management API. All resources require bearer-token authentication unless noted."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API prefix"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error description."
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "detail": {
            "type": "string",
            "description": "Optional extra context."
          }
        }
      },
...
      "SkillBundle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "digest": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "org",
              "team",
              "project",
              "personal"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "deprecated"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceName": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "grants": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "promotions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "Budget": {
        "type": "object",
        "properties": {
          "monthlyLimitUsd": {
            "type": "number"
          },
          "currentSpendUsd": {
            "type": "number"
          },
          "budgetAlertState": {
            "type": "string",
            "enum": [
              "ok",
              "warning",
              "exceeded"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static bearer token. Pass as Authorization: Bearer <token>."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/skills/catalog": {
      "get": {
        "operationId": "listSkillBundles",
        "summary": "List all skill bundles with entitlements and promotion history",
        "tags": [
          "Skills"
        ],
        "responses": {
          "200": {
            "description": "Skill bundle list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillBundle"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createSkillBundle",
        "summary": "Create a new skill bundle",
        "tags": [
          "Skills"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "version",
                  "digest",
                  "scope"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Skill bundle created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/skills/catalog/{id}": {
      "get": {
        "operationId": "getSkillBundle",
        "summary": "Get a single skill bundle by identifier",
        "tags": [
          "Skills"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skill bundle detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillBundle"
                }
              }
            }
          },
          "404": {
            "description": "Skill bundle not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateSkillBundle",
        "summary": "Update a skill bundle and fully replace entitlements and promotions",
        "tags": [
          "Skills"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Skill bundle updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteSkillBundle",
        "summary": "Delete a skill bundle and its linked entitlement grants",
        "tags": [
          "Skills"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skill bundle deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/ai-budget/{tenantName}/spend": {
      "get": {
        "operationId": "getTenantSpend",
        "summary": "Get current spend and budget state for a tenant",
        "tags": [
          "AI Budget"
        ],
        "parameters": [
          {
            "name": "tenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spend data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              }
            }
          },
          "502": {
            "description": "Upstream dependency (Kubernetes, database, Cognee, LiteLLM) returned an error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}